New pricer (#63)
* State of dev before merging master to branch

* Before review

* Code analysis corrections

* Minor revisions after review comments

* Pure virtual AbstractShift and fix some pointer issues

* Initialization of MyRosterSp

- delete Constraint.h
- move resources to a specific folder
- create MyRCGraph, MyRosterSP, MyRCSPPSolver
- instantiate the above
- fix some Clang-tidy issues

* Compute base costs of RCArcs

* Add dual cost and prepare RCSPPsolver

* Add dual costs and prepare RCSPPsolver (everything staged)

* Do not merge? For Windows changes

I had to make those changes, otherwise the code would not compile on Windows. I guess these changes won't hurt, but not sure.
Small addendum to Readme file to talk about Windows install.

* Some todos

* Update MyRCSPP.h

* Update MyRCSPP.cpp

* Changes during meeting JO/AG

* Update MyRosterSP.cpp

* Update MyRCLabel.h

* Update MyRCSPP.h

* Update MyRCSPP.cpp

* Update TotalShiftResource.cpp

* Update ConsShiftResource.cpp

* Update TotalWeekendsResource.cpp

* Update TotalWeekendsResource.h

* Update MyRCSPP.cpp

* Update TotalWeekendsResource.h

* Update MyRCLabel.h

* Meeting with AG

* Refactoring to remove boost and PrincipalGraph from SubProble

* RCSPP solver

* Correction of minor bugs

* before meeting JO/AG

* Minor changes

* Mods meeting JO/AG (1)

* Some TODOs and renamings

* Correct expansion of total weekend and cost of incomplete weekends

* Preprocessing graph

* Improved domination (first version)

* Correction bug in comparator

* new generation of random dual costs

* Correction of minor bug

* Improvements after profiling

* Minor changes

* Start implementing a pool of labels

* Label pools implemented and other cpu optimizations

* Improvements of CPU ok

* Remove ResourceLabel

* Update MyRCLabel.cpp

* Update MyRCSPP.h

* Revert "Update MyRCSPP.h"

This reverts commit c2342734957603a8537358f8aef6330289412f70.

* Enumeration of sub-paths (version 1)

* Minor modifications

* Switch timer to chrono header

It has not been validated on Mac

* Minor modifications

* Enumeration of sub paths (final version) + Correction of minor bugs

* adding RCSPPSolver's options in paramfile

* Minor changes

* New Readme for the tests with the new pricer

* Fix system_clock in Tools as well as inefficient calls to nodes and arcs in MyRCSSP and RCGraph.

* Optimized the MMyRCSPP code. Especially changed the label pool.

* Finished RCSSP optimization.

* Change minor things in ressources definitions.

* Cosmetic changes on MyRoster.

* Create a struct to storethe resources values.

* remove ResourceValues copy when expanding.

* Removed inactive expander.

* Move the computation of worst costs in the resources.

* Fix ResourceValues default consumption as not initialized for the 1rst label. Highly optimized code.

* Replace Shift operator >= and some minor revisions

* Revert "Replace Shift operator >= and some minor revisions"

This reverts commit cfb012691beb576ad8b708187ddc6be0bd463312.

* Revert "Revert "Replace Shift operator >= and some minor revisions""

This reverts commit 0fed270f9fbcbfb748086f602237035ee7a26cc5.

* Small bug corrected

Ok, we get the expected results

* Revise PricerMain to make it a test function without inputs

* Minor changes while progressing in the pull request

* Improved outputs of the main pricer

* Modified the domination loop for sorted option

* Small revision/clarification/optimization for minimumCostToSinksOption_

* Minor revisions while reviewing the enumeration of subpaths

* Not much

* Check if had constraints are satisfied

* Check hard resources when expanding.

* Fix the computation of the cost in the roster boost subproblem: override wasn't correctly changed.

* Add an option to choose SP solver type.

* Fix several bugs in MyRoster as it wasn't debug for reuse. Now works with arg --rcspp-type MY_SOLVER.

* Fix timer for dynamic scheduler. Change pricer main to add some tests in the workflows.

* code cpplint free.

* Removed dual costs for weekend in the MyRoster. Also count the preprocessing time in the solving time of the rcspp.

+++ src/data/Nurse.cpp
@@ -9,7 +9,7 @@
  * full license detail.
  */
 
-#include "Nurse.h"
+#include "data/Nurse.h"
 
 #include <math.h>
 #include <time.h>
@@ -328,7 +328,7 @@ bool Preferences::wantsTheDayOff(int nurseNum, int day) const {
     return false;
   // Set does not repeat its elements. Wants the day off if and only if
   // all shifts off (-1 because REST does not appear)
-  return itM->second.size() == nbShifts_ - 1;
+  return static_cast<int>(itM->second.size()) == nbShifts_ - 1;
 }
 
 // Total number of shifts off that the nurse wants
@@ -397,7 +397,7 @@ bool Preferences::wantsTheDayOn(int nurseNum, int day) const {
     return false;
   // Set does not repeat its elements. Wants the day on if and only if
   // all shifts off (-1 because REST does not appear)
-  return itM->second.size() == nbShifts_ - 1;
+  return static_cast<int>(itM->second.size()) == nbShifts_ - 1;
 }
 
 // Total number of shifts on that the nurse wants
@@ -417,7 +417,7 @@ int Preferences::howManyDaysOn(int nurseNum, int dayMin, int dayMax) const {
   // look at every wishes of the nurse
   for (const auto &dayOn : wishesOn_.at(nurseNum)) {
     nbDayOn += ((dayOn.first >= dayMin) && (dayOn.first <= dayMax)
-        && (dayOn.second.size() == nbShifts_ - 1));
+        && (static_cast<int>(dayOn.second.size()) == nbShifts_ - 1));
   }
   return nbDayOn;
 }
